Fixed com.intellij.util.IncorrectOperationException: Cannot modify compiled element (EA-81990 - IOE: ClsElementImpl.delete)

+++ android/src/org/jetbrains/android/inspections/lint/ParcelableQuickFix.java
@@ -15,6 +15,7 @@
  */
 package org.jetbrains.android.inspections.lint;
 
+import com.intellij.codeInsight.FileModificationService;
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.fileEditor.FileDocumentManager;
 import com.intellij.openapi.project.Project;
@@ -483,7 +484,7 @@ public class ParcelableQuickFix implements AndroidLintQuickFix {
     }
 
     private static void delete(@Nullable PsiElement element) {
-      if (element != null) {
+      if (element != null && FileModificationService.getInstance().preparePsiElementForWrite(element)) {
         element.delete();
       }
     }
